Social: Replace ConnectionIcon component with the implementation of ConnectionImage component.
