+++ projects/js-packages/publicize-components/changelog/update-social-replace-connection-icon-component
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+Social: Replace ConnectionIcon component with the implementation of ConnectionImage component.
